Refactor GUI Map as TileMap
